requirements.txt
flask
Flask-SQLAlchemy
requests
pycrypto
gunicorn